+++ components/block/BlockDetails.tsx
@@ -23,7 +23,6 @@ const useStyles = makeStyles(() => {
   return {
     root: {
       width: "100%",
-      padding: "1%",
       borderRadius: 5,
       wordWrap: "break-word",
     },
@@ -51,6 +50,7 @@ const useStyles = makeStyles(() => {
     iconButtonLeft: {
       padding: "0",
       float: "left",
+      marginLeft: "-0.625rem"
     },
   };
 });
@@ -73,9 +73,9 @@ const BlockDetails = () => {
   return (
     <Card className={classes.root}>
       <CardContent>
-        <Grid container spacing={1} justify="center" className={classes.item}>
+        <Grid container spacing={2} justify="center" className={classes.item}>
           <Grid item xs={10}>
-            <Typography color="textSecondary" variant="subtitle1" paragraph>
+            <Typography color="textPrimary" variant="subtitle1" >
               Block #{number}
             </Typography>
           </Grid>