+++ pages/validatorVotes.tsx
@@ -13,10 +13,6 @@ const useStyles = makeStyles((theme: Theme) =>
             justifyContent: "center",
         },
 
-        bottomPadding: {
-            overflow: "auto",
-            padding: "1rem",
-        },
     })
 );
 
@@ -27,11 +23,11 @@ export default function ValidatorVotes() {
 
     return (
         <Layout>
-            <Grid container className={classes.root}>
-                <Grid item xs={12} sm={8} className={classes.bottomPadding}>
+            <Grid container spacing={2} className={classes.root}>
+                <Grid item xs={12} sm={8} >
                     <MarketCard />
                 </Grid>
-                <Grid item xs={12} sm={8} className={classes.bottomPadding}>
+                <Grid item xs={12} sm={8} >
                     <ValidatorVotesList />
                 </Grid>